Move dashboard migrations into separate module
