+++ public/js/lib/configuration.js
@@ -57,34 +57,3 @@ export function saveDashboard(dashboard) {
 
   return post("/api/dashboard", init);
 }
-
-export function upgradeDashboard(dashboard, migrations = dashboardMigrations) {
-  const storedVersion = dashboard.version || 0;
-  const currentVersion = migrations.length;
-
-  for (let i = storedVersion; i < currentVersion; ++i) {
-    migrations[i](dashboard);
-  }
-
-  dashboard.version = currentVersion;
-  return storedVersion < currentVersion;
-}
-
-const dashboardMigrations = [
-  /**
-   * Dashboard version 1:
-   *
-   * - Number tile: added text color gradient properties
-   */
-  dashboard => {
-    for (const tile of dashboard.tiles) {
-      if (tile.type === "number") {
-        tile.textColorMode = "single";
-        tile.lowValue = 20.0;
-        tile.lowTextColor = "#983b21ff";
-        tile.highValue = 40.0;
-        tile.highTextColor = "#56823fff";
-      }
-    }
-  }
-];
