+++ public/js/lib/specs/dashboardMigrations.spec.js
@@ -1,4 +1,4 @@
-import * as configuration from "../configuration";
+import { upgradeDashboard } from "../dashboardMigrations";
 
 describe("upgradeDashboard", () => {
   it("does not invoke migrations on an up-to-date dashboard", () => {
@@ -6,7 +6,7 @@ describe("upgradeDashboard", () => {
     const migrations = [migration, migration, migration];
     const dashboard = { version: 3 };
 
-    configuration.upgradeDashboard(dashboard, migrations);
+    upgradeDashboard(dashboard, migrations);
 
     expect(migration).not.toHaveBeenCalled();
   });
@@ -19,7 +19,7 @@ describe("upgradeDashboard", () => {
     const migrations = [migration1, migration2, migration3];
     const dashboard = {};
 
-    configuration.upgradeDashboard(dashboard, migrations);
+    upgradeDashboard(dashboard, migrations);
 
     expect(migrationsPerformed).toEqual([1, 2, 3]);
     expect(dashboard.version).toEqual(3);
@@ -33,7 +33,7 @@ describe("upgradeDashboard", () => {
     const migrations = [migration1, migration2, migration3];
     const dashboard = { version: 2 };
 
-    configuration.upgradeDashboard(dashboard, migrations);
+    upgradeDashboard(dashboard, migrations);
 
     expect(migrationsPerformed).toEqual([3]);
     expect(dashboard.version).toEqual(3);
